Added rounding.rs for performing rejection sampling

+++ src/params.rs
@@ -1,12 +1,12 @@
 #![allow(dead_code)]
-pub const M: u8 = 6;
-pub const N: u8 = 128;
+pub const M: usize = 6;
+pub const N: usize = 128;
 pub const Q: i32 = 1073479681;
 pub const GAMMA1: i32 = 1<<18;
 pub const GAMMA2: i32 = (Q-1)/(1<<13);
-pub const D: u8 = 14;
-pub const BETA: u8 = 32;
-pub const R: u8 = 4;
-pub const K: u8 = 10;
-pub const L: u8 = 10;
-pub const SYMBYTES: u8 = 32;
+pub const D: usize = 14;
+pub const BETA: usize = 32;
+pub const R: usize = 4;
+pub const K: usize = 10;
+pub const L: usize = 10;
+pub const SYMBYTES: usize = 32;
